Moved other types to shared module
