--- android/app/src/main/kotlin/pl/kwasow/data/types/AuthenticationResult.kt
@@ -1,12 +1,0 @@
-package pl.kwasow.data.types
-
-data class AuthenticationResult(
-    val authorization: Authorization,
-    val authenticatedUser: User?,
-) {
-    enum class Authorization {
-        AUTHORIZED,
-        UNAUTHORIZED,
-        UNKNOWN,
-    }
-}
