+++ android/app/src/main/kotlin/pl/kwasow/data/types/UserIconDetails.kt
@@ -1,0 +1,9 @@
+package pl.kwasow.data.types
+
+import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
+
+data class UserIconDetails(
+    @DrawableRes val res: Int,
+    @StringRes val description: Int,
+)
