+++ android/app/src/main/kotlin/pl/kwasow/ui/screens/modules/memories/MemoryView.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import pl.kwasow.R
-import pl.kwasow.flamingo.types.Memory
+import pl.kwasow.flamingo.types.memories.Memory
 import pl.kwasow.ui.components.PhotoView
 
 // ====== Public composables
