+++ android/app/src/main/kotlin/pl/kwasow/ui/screens/modules/music/AlbumListView.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.koin.androidx.compose.koinViewModel
 import pl.kwasow.R
-import pl.kwasow.data.types.Album
+import pl.kwasow.flamingo.types.music.Album
 import pl.kwasow.ui.components.PhotoView
 import pl.kwasow.ui.composition.LocalFlamingoNavigation
