+++ android/app/src/main/kotlin/pl/kwasow/ui/screens/modules/music/AlbumTrackList.kt
@@ -32,8 +32,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.koin.androidx.compose.koinViewModel
 import pl.kwasow.R
-import pl.kwasow.data.types.Album
-import pl.kwasow.data.types.AudioTrack
+import pl.kwasow.flamingo.types.music.Album
+import pl.kwasow.flamingo.types.music.AudioTrack
 
 // ====== Public composables
 @Composable
