+++ android/app/src/main/kotlin/pl/kwasow/ui/screens/modules/whishlist/DeleteWishDialog.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import org.koin.androidx.compose.koinViewModel
 import pl.kwasow.R
-import pl.kwasow.data.types.Wish
+import pl.kwasow.flamingo.types.wishlist.Wish
 
 // ====== Public composables
 @Composable
