+++ android/app/src/main/kotlin/pl/kwasow/ui/screens/modules/whishlist/WishlistModuleViewModel.kt
@@ -11,9 +11,9 @@ import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import pl.kwasow.R
-import pl.kwasow.data.types.MinimalUser
 import pl.kwasow.data.types.TabItem
-import pl.kwasow.data.types.Wish
+import pl.kwasow.flamingo.types.user.MinimalUser
+import pl.kwasow.flamingo.types.wishlist.Wish
 import pl.kwasow.managers.UserManager
 import pl.kwasow.managers.WishlistManager
