+++ android/app/src/main/kotlin/pl/kwasow/ui/widgets/memories/MemoryCarousel.kt
@@ -34,7 +34,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import pl.kwasow.R
-import pl.kwasow.flamingo.types.Memory
+import pl.kwasow.flamingo.types.memories.Memory
 
 // ====== Public composables
 @Composable
