+++ android/app/src/main/kotlin/pl/kwasow/ui/widgets/memories/MemoryView.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import pl.kwasow.R
 import pl.kwasow.extensions.nonScaledSp
-import pl.kwasow.flamingo.types.Memory
+import pl.kwasow.flamingo.types.memories.Memory
 import pl.kwasow.ui.components.PhotoView
 
 // ====== Public composables
