+++ android/types/src/main/java/pl/kwasow/flamingo/types/auth/AuthenticationResult.kt
@@ -1,0 +1,8 @@
+package pl.kwasow.flamingo.types.auth
+
+import pl.kwasow.flamingo.types.user.User
+
+data class AuthenticationResult(
+    val authorization: Authorization,
+    val authenticatedUser: User?,
+)
