+++ android/types/src/main/java/pl/kwasow/flamingo/types/auth/Authorization.kt
@@ -1,0 +1,7 @@
+package pl.kwasow.flamingo.types.auth
+
+enum class Authorization {
+    AUTHORIZED,
+    UNAUTHORIZED,
+    UNKNOWN,
+}
