+++ android/types/src/main/java/pl/kwasow/flamingo/types/user/Partner.kt
@@ -1,13 +1,10 @@
-package pl.kwasow.data.types
+package pl.kwasow.flamingo.types.user
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class User(
+data class Partner(
     override val id: Int,
     override val firstName: String,
-    val email: String,
     override val icon: UserIcon,
-    val partner: Partner,
-    val coupleDetails: CoupleDetails,
 ) : MinimalUser
