+++ android/types/src/main/java/pl/kwasow/flamingo/types/user/User.kt
@@ -1,10 +1,13 @@
-package pl.kwasow.data.types
+package pl.kwasow.flamingo.types.user
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Partner(
+data class User(
     override val id: Int,
     override val firstName: String,
+    val email: String,
     override val icon: UserIcon,
+    val partner: Partner,
+    val coupleDetails: CoupleDetails,
 ) : MinimalUser
