+++ android/types/src/main/java/pl/kwasow/flamingo/types/user/UserIcon.kt
@@ -1,0 +1,13 @@
+package pl.kwasow.flamingo.types.user
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class UserIcon {
+    @SerialName("cat")
+    CAT,
+
+    @SerialName("sheep")
+    SHEEP,
+}
